scripts/sim-agent/requirements.txt
httpx>=0.27,<1.0
mcp>=1.0,<2.0
pyyaml>=6.0,<7.0
rich>=13.0,<14.0
python-dotenv>=1.0,<2.0
